requirements.txt
streamlit
tensorflow
pillow
numpy
fastapi
uvicorn
requests
python-multipart
